api/inference-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
torch==2.5.1
torchvision==0.20.1
grad-cam==1.5.5
Pillow==11.0.0
numpy==1.26.4
python-multipart==0.0.12
httpx==0.27.0
